Quick Sort

+++ Lab3/Lab3.cpp
@@ -6,7 +6,7 @@
 int main()
 {
 
-	ArrayList<int> list;
+	ArrayList<double> list;
 	double a[] = { 5.1, 4.9, 4.7, 4.6, 5.0, 5.4, 4.6, 5.0, 4.4, 4.9, 5.4, 4.8, 4.8, 4.3, 5.8, 5.7, 5.4, 5.1, 5.7, 5.1};
 
 	for (int i = 0; i < 20; i++)
